+++ docs/pages/documentation/general/_General.vue
@@ -4,5 +4,6 @@
         <router-link to="icon" class="nav-item is-tab">Icon</router-link>
         <router-link to="dropdown" class="nav-item is-tab">Dropdown</router-link>
         <router-link to="tooltip" class="nav-item is-tab">Tooltip</router-link>
+        <router-link to="modal" class="nav-item is-tab">Modal</router-link>
     </div>
 </template>